Fix bug in handling of `h` literal suffix.
A literal like:

    2.0h

is supposed to default to `half` precision, but there was a typo in our logic that meant that the `half` case was mistakenly checking for `l` and not `h` and so it would never trigger (because the `double` case right before it was checking for `l` first), and certainly wouldn't trigger on an `h`.

There was also a bug that a literal with `hf`:

    2.0hf

would go into the path for the `f` suffix without considering the `h`.

These changes really ought to have some tests for them, but this was also just a quick issue I noticed while working on something else.