Fix declaration of scalar sincos() function.
The function was accidentally defined with a generic `int` parameter copy-pasted  from the vector definition, but that made the scalar version impossible to call with inferred generic arguments, because there wasn't a way to infer `N` when it isn't used in the parameter list.

Includes a simple test case to confirm that the front-end no longer chokes on calls to scalar `sincos()`.